+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* 'Start' page pattern
 * `bsk_footer_classes` config option for adding extra classes to standard page footer
 
 ### Fixed